requirements.txt
requests
tabulate
psycopg2-binary
dataclasses
python-dotenv
fastapi
inquirerPy
argon2
argon2-cffi
Flask
